+++ json_serializable/lib/src/type_helpers/config_types.dart
@@ -11,9 +11,11 @@ class KeyConfig {
 
   final bool disallowNullValue;
 
-  final bool ignore;
+  final bool? includeFromJson;
 
   final bool includeIfNull;
+
+  final bool? includeToJson;
 
   final String name;
 
@@ -26,8 +28,9 @@ class KeyConfig {
   KeyConfig({
     required this.defaultValue,
     required this.disallowNullValue,
-    required this.ignore,
+    required this.includeFromJson,
     required this.includeIfNull,
+    required this.includeToJson,
     required this.name,
     required this.readValueFunctionName,
     required this.required,
@@ -46,6 +49,7 @@ class ClassConfig {
   final bool createFactory;
   final bool createToJson;
   final bool createFieldMap;
+  final bool createPerFieldToJson;
   final bool disallowUnrecognizedKeys;
   final bool explicitToJson;
   final FieldRename fieldRename;
@@ -62,6 +66,7 @@ class ClassConfig {
     required this.createFactory,
     required this.createToJson,
     required this.createFieldMap,
+    required this.createPerFieldToJson,
     required this.disallowUnrecognizedKeys,
     required this.explicitToJson,
     required this.fieldRename,
@@ -80,6 +85,8 @@ class ClassConfig {
         constructor: config.constructor ?? ClassConfig.defaults.constructor,
         createFieldMap:
             config.createFieldMap ?? ClassConfig.defaults.createFieldMap,
+        createPerFieldToJson: config.createPerFieldToJson ??
+            ClassConfig.defaults.createPerFieldToJson,
         createFactory:
             config.createFactory ?? ClassConfig.defaults.createFactory,
         createToJson: config.createToJson ?? ClassConfig.defaults.createToJson,
@@ -106,6 +113,7 @@ class ClassConfig {
     createFactory: true,
     createToJson: true,
     createFieldMap: false,
+    createPerFieldToJson: false,
     disallowUnrecognizedKeys: false,
     explicitToJson: false,
     fieldRename: FieldRename.none,
@@ -121,6 +129,7 @@ class ClassConfig {
         createFactory: createFactory,
         createToJson: createToJson,
         createFieldMap: createFieldMap,
+        createPerFieldToJson: createPerFieldToJson,
         ignoreUnannotated: ignoreUnannotated,
         explicitToJson: explicitToJson,
         includeIfNull: includeIfNull,