+++ json_serializable/lib/src/type_helpers/to_from_string.dart
@@ -5,7 +5,7 @@
 import 'package:analyzer/dart/element/type.dart';
 import 'package:source_gen/source_gen.dart';
 
-import '../utils.dart';
+import '../default_container.dart';
 
 final bigIntString = ToFromStringHelper(
   'BigInt.parse',
@@ -67,7 +67,7 @@ class ToFromStringHelper {
     return '$expression.$_toString';
   }
 
-  String? deserialize(
+  DefaultContainer? deserialize(
     DartType type,
     String expression,
     bool nullable,
@@ -84,6 +84,9 @@ class ToFromStringHelper {
       output = '$output$_parseSuffix';
     }
 
-    return nullable ? ifNullOrElse(expression, 'null', output) : output;
+    return DefaultContainer(
+      expression,
+      output,
+    );
   }
 }