+++ json_serializable/test/integration/json_enum_example.dart
@@ -25,6 +25,36 @@ enum DayType {
 
 Iterable<String> get dayTypeEnumValues => _$DayTypeEnumMap.values;
 
+@JsonEnum(alwaysCreate: true, valueField: 'value')
+enum MyStatusCode {
+  success(200),
+  @JsonValue(701) // explicit value always takes precedence
+  weird(601);
+
+  const MyStatusCode(this.value);
+
+  final int value;
+}
+
+Iterable<int> get myStatusCodeEnumValues => _$MyStatusCodeEnumMap.values;
+
+@JsonEnum(alwaysCreate: true, valueField: 'index')
+enum EnumValueFieldIndex {
+  success(200),
+  @JsonValue(701) // explicit value always takes precedence
+  weird(601),
+  oneMore(777);
+
+  static const tryingToBeConfusing = weird;
+
+  const EnumValueFieldIndex(this.value);
+
+  final int value;
+}
+
+Iterable<int> get enumValueFieldIndexValues =>
+    _$EnumValueFieldIndexEnumMap.values;
+
 @JsonSerializable(
   createToJson: false,
 )
@@ -73,11 +103,21 @@ class Issue1145RegressionA {
   createFactory: false,
 )
 class Issue1145RegressionB {
-  Issue1145RegressionB({
-    required this.status,
-  });
+  Issue1145RegressionB({required this.status});
 
   Map<String, dynamic> toJson() => _$Issue1145RegressionBToJson(this);
 
   final List<Issue1145RegressionEnum?> status;
+}
+
+@JsonSerializable(includeIfNull: false)
+class Issue1226Regression {
+  Issue1226Regression({required this.durationType});
+
+  factory Issue1226Regression.fromJson(Map<String, dynamic> json) =>
+      _$Issue1226RegressionFromJson(json);
+
+  final Issue1145RegressionEnum? durationType;
+
+  Map<String, dynamic> toJson() => _$Issue1226RegressionToJson(this);
 }